requirements.txt
python==3.6.7
numpy==1.13.3
pytorch==0.4.1
statsmodels==0.9.0
shapely==1.6.4
six==1.11.0
wheel=0.31.1
scipy==0.19.1
scikit-learn==0.18.1
scikit-image==0.14.0
pywavelets==0.5.2
py-opencv==3.4.2
pillow==5.2.0
pandas==0.23.2
opencv==3.4.2
matplotlib==2.2.3
Cmake>=3.10
Ninja
svgpathtools
svgutils
svgwrite
